Separate command out from main & export as lib
