+++ src/lib.rs
@@ -1,0 +1,3 @@
+mod command;
+
+pub use command::Command;
